Move portal container context to ds-components
