+++ libs/@hashintel/petrinaut/.storybook/preview.tsx
@@ -1,7 +1,7 @@
 import "../src/ui/index.css";
 import { useRef } from "react";
 
-import { PortalContainerContext } from "../src/react/state/portal-container-context";
+import { PortalContainerContext } from "@hashintel/ds-components";
 
 import type { Preview } from "@storybook/react-vite";
