+++ libs/@hashintel/petrinaut/src/ui/components/drawer.tsx
@@ -9,11 +9,10 @@ import {
 } from "react";
 import { createPortal } from "react-dom";
 
-import { Button } from "@hashintel/ds-components";
+import { Button, PortalContainerContext } from "@hashintel/ds-components";
 import { css, cx } from "@hashintel/ds-helpers/css";
 
 import { createValueStore } from "../../react/create-value-store";
-import { PortalContainerContext } from "../../react/state/portal-container-context";
 import {
   Body,
   Card as PanelCard,
