+++ libs/@hashintel/petrinaut/src/ui/components/menu.tsx
@@ -1,9 +1,8 @@
 import { Menu as ArkMenu } from "@ark-ui/react";
 import { Portal } from "@ark-ui/react/portal";
 
+import { usePortalContainerRef } from "@hashintel/ds-components";
 import { css, cva } from "@hashintel/ds-helpers/css";
-
-import { usePortalContainerRef } from "../../react/state/portal-container-context";
 
 import type { ReactNode } from "react";
