+++ libs/@hashintel/petrinaut/src/ui/components/popover.tsx
@@ -1,10 +1,8 @@
 import { Popover as ArkPopover } from "@ark-ui/react/popover";
 import { Portal } from "@ark-ui/react/portal";
 
-import { Button } from "@hashintel/ds-components";
+import { Button, usePortalContainerRef } from "@hashintel/ds-components";
 import { css, cx } from "@hashintel/ds-helpers/css";
-
-import { usePortalContainerRef } from "../../react/state/portal-container-context";
 
 import type { ComponentProps, HTMLAttributes, ReactNode } from "react";
